Fix setting stickyPointerLock game-shell option default
If pointerLock is enabled, then also enable stickyPointerLock, in order
to continuously re-acquire pointer lock. This is needed to re-acquire after
pointer lock is released, but if undesirable behavior this default can be
overridden with the stickyPointerLock option to voxel-engine-stackgl.

Setting this option fixes a regression - previously (circa 2015/4) it was
not needed, but now it is required for correct behavior or you cannot "re-enter"
the game after exiting (relinquishing pointer lock with Esc). See:

https://github.com/deathcap/voxelmetaverse/issues/65 Pointer lock cannot be re-acquired after releasing
https://github.com/mikolalysenko/game-shell/pull/8 `stickyPointerLock` typos fix